Use promotion copy in the GW landing page
