+++ support-frontend/assets/helpers/globals.js
@@ -1,6 +1,7 @@
 // @flow
 import type { ProductPrices } from 'helpers/productPrice/productPrices';
 import { type Settings } from 'helpers/settings';
+import type { LandingPage } from 'helpers/productPrice/promotions';
 
 function getGlobal<T>(path: string = ''): ?T {
 
@@ -49,6 +50,8 @@ const getSettings = (): Settings => getGlobal('settings') || {
 
 const getProductPrices = (): ?ProductPrices => getGlobal('productPrices');
 
+const getPromotionCopy = (): ?LandingPage => getGlobal('promotionCopy');
+
 const isSwitchOn = (switchName: string): boolean => {
   const sw = getGlobal(`settings.switches.${switchName}`);
   return !!(sw && sw === 'On');
@@ -64,4 +67,4 @@ const isTestSwitchedOn = (testName: string): boolean => {
 };
 
 
-export { getProductPrices, getGlobal, isTestSwitchedOn, getSettings, isSwitchOn };
+export { getProductPrices, getPromotionCopy, getGlobal, isTestSwitchedOn, getSettings, isSwitchOn };
