+++ support-frontend/assets/helpers/productPrice/productPrices.js
@@ -21,41 +21,10 @@ import {
   type IsoCurrency,
 } from 'helpers/internationalisation/currency';
 import { fixDecimals } from 'helpers/subscriptions';
-import type { Option } from 'helpers/types/option';
-import { getQueryParameter } from 'helpers/url';
+import type { Promotion } from 'helpers/productPrice/promotions';
+import { applyDiscount, getPromotion } from 'helpers/productPrice/promotions';
 
 // ----- Types ----- //
-
-export type DiscountBenefit = {
-  amount: number,
-  durationMonths?: number,
-}
-
-export type IntroductoryPeriodType = 'issue';
-
-export type IntroductoryPriceBenefit = {
-  price: number,
-  periodLength: number,
-  periodType: IntroductoryPeriodType,
-}
-
-export type LandingPage = {
-  title?: string,
-  description?: string,
-  roundel?: string,
-}
-
-export type Promotion =
-  {
-    name: string,
-    description: string,
-    promoCode: string,
-    discountedPrice?: number,
-    numberOfDiscountedPeriods?: number,
-    discount?: DiscountBenefit,
-    introductoryPrice?: IntroductoryPriceBenefit,
-    landingPage?: LandingPage,
-  }
 
 export type ProductPrice = {
   price: number,
@@ -86,32 +55,6 @@ const isNumeric = (num: ?number): boolean %checks =>
   num !== undefined &&
   !Number.isNaN(num);
 
-const hasDiscount = (promotion: ?Promotion): boolean %checks =>
-  promotion !== null &&
-  promotion !== undefined &&
-  isNumeric(promotion.discountedPrice);
-
-const hasIntroductoryPrice = (promotion: ?Promotion): boolean %checks =>
-  promotion !== null &&
-  promotion !== undefined &&
-  !!promotion.introductoryPrice;
-
-function applyDiscount(price: ProductPrice, promotion: ?Promotion) {
-  if (hasDiscount(promotion)) {
-    return {
-      ...price,
-      price: promotion.discountedPrice,
-    };
-  } else if (hasIntroductoryPrice(promotion)) {
-    return {
-      ...price,
-      // $FlowIgnore - we have checked this above
-      price: promotion.introductoryPrice.price,
-    };
-  }
-  return price;
-}
-
 function getCountryGroup(country: IsoCountry): CountryGroup {
   return countryGroups[fromCountry(country) || GBPCountries];
 }
@@ -128,41 +71,6 @@ function getProductPrice(
   return productPrices[countryGroup.name][fulfilmentOption ||
   NoFulfilmentOptions][productOption || NoProductOptions][billingPeriod ===
   SixWeekly ? Quarterly : billingPeriod][countryGroup.currency];
-}
-
-const matchesQueryParam = promotion =>
-  getQueryParameter('promoCode') === promotion.promoCode;
-
-const introductoryPrice = promotion =>
-  promotion.introductoryPrice !== null && promotion.introductoryPrice !==
-  undefined;
-
-function getAppliedPromo(promotions: ?Promotion[]): Option<Promotion> {
-  if (promotions && promotions.length > 0) {
-    if (promotions.length > 1) {
-      return promotions.find(introductoryPrice) ||
-        promotions.find(matchesQueryParam) ||
-        promotions[0];
-    }
-    return promotions[0];
-  }
-  return null;
-}
-
-function getPromotion(
-  productPrices: ProductPrices,
-  country: IsoCountry,
-  billingPeriod: BillingPeriod,
-  fulfilmentOption: ?FulfilmentOptions,
-  productOption: ?ProductOptions,
-): ?Promotion {
-  return getAppliedPromo(getProductPrice(
-    productPrices,
-    country,
-    billingPeriod,
-    fulfilmentOption,
-    productOption,
-  ).promotions);
 }
 
 function finalPrice(
@@ -215,16 +123,11 @@ function getCurrency(country: IsoCountry): IsoCurrency {
 }
 
 export {
-  getAppliedPromo,
   getProductPrice,
-  getPromotion,
   finalPrice,
   getCurrency,
   getCountryGroup,
   showPrice,
   displayPrice,
-  applyDiscount,
-  hasDiscount,
-  hasIntroductoryPrice,
   isNumeric,
 };
