+++ support-frontend/assets/pages/weekly-subscription-landing/weeklySubscriptionLandingReducer.js
@@ -4,16 +4,19 @@
 
 import type { CommonState } from 'helpers/page/commonReducer';
 import type { ProductPrices } from 'helpers/productPrice/productPrices';
-import { getProductPrices } from 'helpers/globals';
+import { getProductPrices, getPromotionCopy } from 'helpers/globals';
+import type { LandingPage } from 'helpers/productPrice/promotions';
 
 export type State = {
   common: CommonState,
   page: {
     productPrices: ?ProductPrices,
+    promotionCopy: ?LandingPage,
   }
 };
 
 // ----- Export ----- //
 export default () => ({
   productPrices: getProductPrices(),
+  promotionCopy: getPromotionCopy(),
 });
